livekit-voice-agent/requirements.txt
livekit-agents[silero,turn-detector]~=1.2
livekit-plugins-noise-cancellation~=0.2
livekit-plugins-openai
python-dotenv
